requirements.txt
crewai==0.75
crewai_tools
PyYAML
python-dotenv